requirements.txt
fastapi
uvicorn
requests
python-multipart
Pillow
rapidocr-paddle
opencv-python-headless
numpy
